Package pl.edu.agh.iet.tsp.login.oauth2.configuration renamed to *.config to match conventions.
