+++ tsp-login/src/main/java/pl/edu/agh/iet/tsp/login/oauth2/config/ClientConfiguration.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.iet.tsp.login.oauth2.configuration;
+package pl.edu.agh.iet.tsp.login.oauth2.config;
 
 import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
 import org.springframework.security.oauth2.client.token.grant.code.AuthorizationCodeResourceDetails;
